Update dependencies (including clap)
No functionality changes, but some back-end changes, especially to
clap, which I updated from v2 to v4.  v4 introduced a
whole bunch of breaking changes, but I have been working to resolve
all of these.

While it was a bit difficult figuring it all out (especially the
errors where the compiler doesn't fail but the programme panics, such
as when I tried to `get_one` as `&str` and it gave me the most
unhelpful error message---see
https://stackoverflow.com/a/74344215/12069968), I think the new clap
API is so much cleaner.