Upgrade dependencies and Rust edition!  v2.0.2
Upgrade Rust edition from 2015 to 2024.  Accordingly, apply `cargo
fmt` and `rustfmt` suggestions to codebase.  We also upgrade project
dependencies (minimal breaking changes, only to `base64`).

While cleaning up, we move nested logic in `main` function into other
functions in `main.rs` for readability.  We also pretty print address
results for readability.

Finally, we specify in help that the `--coordinates` flag is not yet
ready to use (it is a work in progress).